[REFACTOR] #13 - presenter/customview -> presenter/custom 으로 수정
